+++ docs/whats_new.rst
@@ -35,6 +35,7 @@ Bugs Fixed
 
 * Fixed typing of :func:`Member.move_to`.
 * Add proper exports for ``discord.ui``.
+* Force autocomplete parameters to be required.
 
 .. _vp0_0_4: